+++ sdtv4/__init__.py
@@ -1,3 +1,4 @@
 #from sdtv4.SDT4PrintOneM2MXSD import print4OneM2MXSD
 from sdtv4.SDT4Parser import SDT4Parser
 from sdtv4.SDT4Classes import *
+from common.SDTHelper import OutType